requirements.txt
aioeos==1.0.2
discord.py==2.3.2
